requirements.txt
google-generativeai
Flask
gTTS
playsound
requests
markupsafe